Separate folders by client version
