+++ netspresso/clients/compressor/__init__.py
@@ -1,3 +1,4 @@
-from .main import compressor_client
+from netspresso.clients.compressor.v1.main import compressor_client
+from netspresso.clients.compressor.v2.main import compressor_client_v2
 
-__all__ = ["compressor_client"]
+__all__ = ["compressor_client", "compressor_client_v2"]
