+++ netspresso/clients/compressor/v1/schemas/compression.py
