+++ netspresso/clients/compressor/v1/schemas/model.py
